fix: Resolve all TypeScript compilation errors
- Fixed admin service type errors:
  * Added explicit type annotation to reduce function
  * Fixed totals object typing
  * Removed invalid Prisma _count include
  * Replaced with manual transaction count queries

- Fixed transaction service type error:
  * Added explicit bigint cast for balance.toString()

All TypeScript errors resolved ✅
Ready for Phase 11 testing

+++ apps/raverpay-api/src/alchemy/transactions/alchemy-transaction.service.ts
@@ -296,7 +296,7 @@ export class AlchemyTransactionService {
       tokenType,
       tokenAddress,
       balance: formattedBalance,
-      balanceRaw: balance.toString(),
+      balanceRaw: (balance as bigint).toString(),
       blockchain: wallet.blockchain,
       network: wallet.network,
     };
